requirements.txt
# ============================================
# Core
# ============================================
python-dotenv==1.0.1
pydantic>=2,<3
numpy==1.26.4
packaging==24.1
urllib3==2.3.0

# ============================================
# LangChain 0.3.x Ecosystem
# ============================================
langchain==0.3.7
langchain-core==0.3.63
langchain-community==0.3.7
langchain-ollama==0.2.0
langchain-google-genai==2.0.4

# ============================================
# Vector Stores
# ============================================
chromadb==0.5.5

# ============================================
# Embeddings / Transformers
# ============================================
transformers==4.44.2
huggingface_hub==0.24.6
sentence-transformers==3.0.1
tiktoken==0.7.0
scikit-learn==1.4.2
scipy==1.11.4

# ============================================
# PDF & file parsing
# ============================================
pypdf==4.3.1
pymupdf==1.24.10

# ============================================
# Audio / Whisper (OpenAI Whisper fork)
# ============================================
openai-whisper==20231117
torch==2.2.2

# ============================================
# Misc tools
# ============================================
python-dateutil==2.9.0
pandas==2.1.4
pyyaml==6.0.2
tqdm==4.66.4